Shrink memory 3: smaller link counts in inode_info

+++ gfs2/fsck/fsck.h
@@ -13,9 +13,9 @@ struct inode_info
 {
         osi_list_t list;
         uint64_t   inode;
-        uint32_t   link_count;   /* the number of links the inode
+        uint16_t   link_count;   /* the number of links the inode
                                   * thinks it has */
-        uint32_t   counted_links; /* the number of links we've found */
+        uint16_t   counted_links; /* the number of links we've found */
 };
 
 struct dir_info
